Swap to a Task SDK implementation
