+++ airflow-core/tests/unit/models/test_deadline.py
@@ -18,14 +18,18 @@ from __future__ import annotations
 
 import json
 import logging
+from datetime import datetime
+from unittest import mock
 
 import pytest
 import time_machine
 from sqlalchemy import select
+from sqlalchemy.exc import SQLAlchemyError
 
 from airflow.models import DagRun
-from airflow.models.deadline import Deadline, DeadlineAlert
+from airflow.models.deadline import Deadline, DeadlineAlert, _fetch_from_db
 from airflow.providers.standard.operators.empty import EmptyOperator
+from airflow.sdk.definitions.deadline_reference import DeadlineReference
 from airflow.utils.state import DagRunState
 
 from tests_common.test_utils import db
@@ -182,3 +186,126 @@ class TestDeadlineAlert:
 
             assert "could not be imported" in caplog.text
             assert path == UNIMPORTABLE_DOT_PATH
+
+
+@pytest.mark.db_test
+class TestCalculatedDeadlineDatabaseCalls:
+    @staticmethod
+    def setup_method():
+        _clean_db()
+
+    @staticmethod
+    def teardown_method():
+        _clean_db()
+
+    @pytest.mark.parametrize(
+        "column, conditions, expected_query",
+        [
+            pytest.param(
+                DagRun.logical_date,
+                {"dag_id": DAG_ID},
+                "SELECT dag_run.logical_date \nFROM dag_run \nWHERE dag_run.dag_id = :dag_id_1",
+                id="single_condition_logical_date",
+            ),
+            pytest.param(
+                DagRun.queued_at,
+                {"dag_id": DAG_ID},
+                "SELECT dag_run.queued_at \nFROM dag_run \nWHERE dag_run.dag_id = :dag_id_1",
+                id="single_condition_queued_at",
+            ),
+            pytest.param(
+                DagRun.logical_date,
+                {"dag_id": DAG_ID, "state": "running"},
+                "SELECT dag_run.logical_date \nFROM dag_run \nWHERE dag_run.dag_id = :dag_id_1 AND dag_run.state = :state_1",
+                id="multiple_conditions",
+            ),
+        ],
+    )
+    @mock.patch("sqlalchemy.orm.Session")
+    def test_fetch_from_db_success(self, mock_session, column, conditions, expected_query):
+        """Test successful database queries."""
+        mock_session.scalar.return_value = DEFAULT_DATE
+
+        result = _fetch_from_db(column, session=mock_session, **conditions)
+
+        assert isinstance(result, datetime)
+        mock_session.scalar.assert_called_once()
+
+        # Check that the correct query was constructed
+        call_args = mock_session.scalar.call_args[0][0]
+        assert str(call_args) == expected_query
+
+        # Verify the actual parameter values
+        compiled = call_args.compile()
+        for key, value in conditions.items():
+            # Note that SQLAlchemy appends the _1 to ensure unique template field names
+            assert compiled.params[f"{key}_1"] == value
+
+    @pytest.mark.parametrize(
+        "use_valid_conditions, scalar_side_effect, expected_error, expected_message",
+        [
+            pytest.param(
+                False,
+                mock.DEFAULT,  # This will allow the call to pass through
+                AttributeError,
+                None,
+                id="invalid_attribute",
+            ),
+            pytest.param(
+                True,
+                SQLAlchemyError("Database connection failed"),
+                SQLAlchemyError,
+                "Database connection failed",
+                id="database_error",
+            ),
+            pytest.param(
+                True, lambda x: None, ValueError, "No matching record found in the database", id="no_results"
+            ),
+        ],
+    )
+    @mock.patch("sqlalchemy.orm.Session")
+    def test_fetch_from_db_error_cases(
+        self, mock_session, use_valid_conditions, scalar_side_effect, expected_error, expected_message, caplog
+    ):
+        """Test database access error handling."""
+        model_reference = DagRun.logical_date
+        conditions = {"dag_id": "test_dag"} if use_valid_conditions else {"non_existent_column": "some_value"}
+
+        # Configure mock session
+        mock_session.scalar.side_effect = scalar_side_effect
+
+        with caplog.at_level(logging.ERROR):
+            with pytest.raises(expected_error, match=expected_message):
+                _fetch_from_db(model_reference, session=mock_session, **conditions)
+            if expected_message:
+                assert expected_message in caplog.text
+
+    @pytest.mark.parametrize(
+        "reference, expected_column",
+        [
+            pytest.param(DeadlineReference.DAGRUN_LOGICAL_DATE, DagRun.logical_date, id="logical_date"),
+            pytest.param(DeadlineReference.DAGRUN_QUEUED_AT, DagRun.queued_at, id="queued_at"),
+            pytest.param(DeadlineReference.FIXED_DATETIME(DEFAULT_DATE), None, id="fixed_deadline"),
+        ],
+    )
+    def test_deadline_database_integration(self, reference, expected_column):
+        """
+        Test database integration for all deadline types.
+
+        Verifies:
+        1. Calculated deadlines call _fetch_from_db with correct column.
+        2. Fixed deadlines do not interact with database.
+        """
+        conditions = {"dag_id": DAG_ID}
+
+        with mock.patch("airflow.sdk.definitions.deadline_reference._fetch_from_db") as mock_fetch:
+            mock_fetch.return_value = DEFAULT_DATE
+
+            if expected_column is not None:
+                result = reference.evaluate_with(**conditions)
+                mock_fetch.assert_called_once_with(expected_column, **conditions)
+            else:
+                result = reference.evaluate_with(**conditions)
+                mock_fetch.assert_not_called()
+
+            assert result == DEFAULT_DATE
